Initial implementation of filteroperator computation: Still needs i) parallel search, ii) computation of weights iii) application of filter to sensitivities. Basically just computes local graph currently.

+++ src/ATO/utils/ATO_Optimizer.hpp
@@ -45,7 +45,7 @@ class Optimizer
   Teuchos::RCP<const Epetra_Comm> comm;
 
   double _optConvTol;
-  double _optMaxIter;
+  int    _optMaxIter;
 
 };
